+++ config/urls.py
@@ -1,3 +1,4 @@
+from django.conf import settings
 from django.contrib import admin
 from django.urls import include, path
 
@@ -14,4 +15,10 @@ urlpatterns = [
     path("api/", include("lemarche.api.urls")),
     path("", include("lemarche.www.home.urls")),
     path("", include("lemarche.www.pages.urls")),
+    path("", include("django.contrib.flatpages.urls")),
 ]
+
+if settings.DEBUG and "debug_toolbar" in settings.INSTALLED_APPS:
+    import debug_toolbar
+
+    urlpatterns = [path("__debug__/", include(debug_toolbar.urls))] + urlpatterns