+++ lemarche/cocorico/management/commands/migrate_data_to_django.py
@@ -10,9 +10,10 @@ from django.db.models.fields import BooleanField, DateTimeField
 from django.utils import timezone
 from django.utils.text import slugify
 
+from lemarche.siaes.models import Siae, SiaeOffer, SiaeLabel, SiaeClientReference
 from lemarche.networks.models import Network
-from lemarche.sectors.models import Sector, SectorGroup
-from lemarche.siaes.models import Siae, SiaeLabel, SiaeOffer
+from lemarche.sectors.models import SectorGroup, Sector
+from lemarche.users.models import User
 
 
 DIRECTORY_EXTRA_KEYS = [
@@ -23,11 +24,48 @@ DIRECTORY_EXTRA_KEYS = [
     "sector",  # string 'list' with ' - ' seperator. We can map to Sector.
     # But we use instead the 'directory_category' table.
 ]
+USER_EXTRA_KEYS = [
+    "username",
+    "username_canonical",
+    "email_canonical",
+    "slug",
+    "salt",
+    "password",
+    "confirmation_token",
+    "password_requested_at",
+    "roles",
+    "person_type",
+    "birthday",
+    "nationality",
+    "country_of_residence",
+    "profession",
+    "mother_tongue",
+    # "phone_prefix", "time_zone", "phone_verified", "email_verified", "id_card_verified",
+    # "accept_survey", "accept_rgpd", "offers_for_pro_sector", "quote_promise",
+    "iban",
+    "bic",
+    "bank_owner_name",
+    "bank_owner_address",
+    "annual_income",
+    "nb_bookings_offerer",
+    "nb_bookings_asker",
+    "fee_as_asker",
+    "fee_as_offerer",
+    "average_rating_as_asker",
+    "average_rating_as_offerer",
+    "answer_delay",
+    "nb_quotes_offerer",
+    "nb_quotes_asker",
+    "company_addr_string",
+]
 
 DIRECTORY_BOOLEAN_FIELDS = [field.name for field in Siae._meta.fields if type(field) == BooleanField]
+USER_BOOLEAN_FIELDS = [field.name for field in User._meta.fields if type(field) == BooleanField]
+
 DIRECTORY_DATE_FIELDS = [field.name for field in Siae._meta.fields if type(field) == DateTimeField]
 NETWORK_DATE_FIELDS = [field.name for field in Network._meta.fields if type(field) == DateTimeField]
 SECTOR_DATE_FIELDS = [field.name for field in Sector._meta.fields if type(field) == DateTimeField]
+USER_DATE_FIELDS = [field.name for field in User._meta.fields if type(field) == DateTimeField]
 
 
 def dsn2params(dsn):
@@ -39,12 +77,20 @@ def dsn2params(dsn):
     return d
 
 
-def integer_to_boolean(input_value):
-    if input_value in [1]:
-        return True
-    elif input_value in [0, None]:
-        return False
-    return False
+def rename_field(elem, field_name_before, field_name_after):
+    elem[field_name_after] = elem[field_name_before]
+    elem.pop(field_name_before)
+
+
+def integer_to_boolean(elem):
+    boolean_keys = list(set(DIRECTORY_BOOLEAN_FIELDS + USER_BOOLEAN_FIELDS))
+    for key in boolean_keys:
+        if key in elem:
+            if elem[key] in [1]:
+                elem[key] = True
+            elif elem[key] in [0, None]:
+                elem[key] = False
+            elem[key] = False
 
 
 def cleanup_date_field_names(elem):
@@ -59,7 +105,7 @@ def cleanup_date_field_names(elem):
 
 
 def make_aware_dates(elem):
-    date_keys = list(set(DIRECTORY_DATE_FIELDS + NETWORK_DATE_FIELDS + SECTOR_DATE_FIELDS))
+    date_keys = list(set(DIRECTORY_DATE_FIELDS + NETWORK_DATE_FIELDS + SECTOR_DATE_FIELDS + USER_DATE_FIELDS))
     for key in date_keys:
         if key in elem:
             if elem[key]:
@@ -80,7 +126,28 @@ def map_presta_type(input_value_byte):
             "12": [Siae.PRESTA_PREST, Siae.PRESTA_BUILD],
             "14": [Siae.PRESTA_DISP, Siae.PRESTA_PREST, Siae.PRESTA_BUILD],
         }
-        return presta_type_mapping[input_value_string]
+        try:
+            return presta_type_mapping[input_value_string]
+        except:  # noqa
+            pass
+    return None
+
+
+def map_user_kind(input_value_integer):
+    if input_value_integer:
+        user_kind_mapping = {
+            None: None,
+            # 1: User.KIND_PERSO,
+            # 2: User.KIND_COMPANY,
+            3: User.KIND_BUYER,
+            4: User.KIND_SIAE,
+            5: User.KIND_ADMIN,
+            6: User.KIND_PARTNER,
+        }
+        try:
+            return user_kind_mapping[input_value_integer]
+        except:  # noqa
+            pass
     return None
 
 
@@ -97,7 +164,7 @@ def reset_app_sql_sequences(app_name):
     with connection.cursor() as cursor:
         cursor.execute(sql)
     output.close()
-    print("Done !")
+    print("Reset complete !")
 
 
 class Command(BaseCommand):
@@ -109,6 +176,12 @@ class Command(BaseCommand):
     directory_network --> M2M between Siae & Network
     listing_category & listing_category_translation --> Sector
     directory_listing_category --> M2M between Siae & Sector
+    directory_label --> SiaeLabel ("Labels & certifications") + OneToMany between Siae & Label
+    directory_offer --> SiaeOffer ("Prestations proposées") + OneToMany between Siae & Offer
+    directory_client_image --> SiaeClientReference ("Références clients") +
+        OneToMany between Siae & SiaeClientReference
+    user --> User
+    directory_user --> M2M between Siae & User
 
     Usage: poetry run python manage.py migrate_data_to_django
     """
@@ -127,6 +200,9 @@ class Command(BaseCommand):
                 self.migrate_siae_sector(cur)
                 self.migrate_siae_offer(cur)
                 self.migrate_siae_label(cur)
+                self.migrate_siae_client_reference(cur)
+                self.migrate_user(cur)
+                self.migrate_siae_user(cur)
         except Exception as e:
             # logger.exception(e)
             print(e)
@@ -153,14 +229,10 @@ class Command(BaseCommand):
         # print(elem)
 
         for elem in resp:
-            # cleanup boolean fields
-            for key in DIRECTORY_BOOLEAN_FIELDS:
-                if key in elem:
-                    elem[key] = integer_to_boolean(elem[key])
-
-            # cleanup dates
+            # cleanup fields
             cleanup_date_field_names(elem)
             make_aware_dates(elem)
+            integer_to_boolean(elem)
 
             # cleanup presta_type
             if "presta_type" in elem:
@@ -172,7 +244,6 @@ class Command(BaseCommand):
             # create object
             try:
                 first = Siae.objects.create(**elem)  # noqa
-                # print(first.__dict__)
             except Exception as e:
                 print(e)
 
@@ -350,15 +421,12 @@ class Command(BaseCommand):
         # print(elem)
 
         for elem in resp:
-            # cleanup dates
+            # rename fields
+            rename_field(elem, "directory_id", "siae_id")
+
+            # cleanup fields
             cleanup_date_field_names(elem)
             make_aware_dates(elem)
-
-            # cleanup relation
-            elem["siae_id"] = elem["directory_id"]
-
-            # remove useless keys
-            [elem.pop(key) for key in ["directory_id"]]
 
             # create object
             SiaeOffer.objects.create(**elem)
@@ -366,7 +434,9 @@ class Command(BaseCommand):
         print(f"Created {SiaeOffer.objects.count()} offers !")
 
     def migrate_siae_label(self, cur):
-        """ """
+        """
+        Migrate SiaeLabel data
+        """
         print("Migrating SiaeLabel...")
 
         SiaeLabel.objects.all().delete()
@@ -383,17 +453,141 @@ class Command(BaseCommand):
         # print(elem)
 
         for elem in resp:
-            # cleanup dates
+            # rename fields
+            rename_field(elem, "directory_id", "siae_id")
+
+            # cleanup fields
             cleanup_date_field_names(elem)
             make_aware_dates(elem)
-
-            # cleanup relation
-            elem["siae_id"] = elem["directory_id"]
-
-            # remove useless keys
-            [elem.pop(key) for key in ["directory_id"]]
 
             # create object
             SiaeLabel.objects.create(**elem)
 
         print(f"Created {SiaeLabel.objects.count()} labels !")
+
+    def migrate_siae_client_reference(self, cur):
+        """
+        Migrate SiaeClientReference data
+        """
+        print("Migrating SiaeClientReference...")
+
+        SiaeClientReference.objects.all().delete()
+
+        cur.execute("SELECT * FROM directory_client_image")
+        resp = cur.fetchall()
+        # print(len(resp))
+        # print(resp[0])
+
+        # l = [elem["position"] for elem in resp]
+        # print(Counter(l))
+
+        # elem = cur.fetchone()
+        # print(elem)
+
+        for elem in resp:
+            # cleanup dates
+            cleanup_date_field_names(elem)
+            make_aware_dates(elem)
+
+            # rename fields
+            rename_field(elem, "name", "image_name")
+            rename_field(elem, "description", "name")
+            rename_field(elem, "position", "order")
+            rename_field(elem, "directory_id", "siae_id")
+
+            # create object
+            SiaeClientReference.objects.create(**elem)
+
+        print(f"Created {SiaeClientReference.objects.count()} client references !")
+
+    def migrate_user(self, cur):
+        """
+        Migrate User data
+        """
+        print("Migrating User...")
+
+        User.objects.filter(api_key__isnull=True).delete()
+        reset_app_sql_sequences("users")
+
+        cur.execute("SELECT * FROM user")
+        resp = cur.fetchall()
+        # print(len(resp))
+
+        # s = set([elem["answer_delay"] for elem in resp])
+        # print(s)
+
+        # elem = cur.fetchone()
+        # print(elem)
+
+        for elem in resp:
+            # rename fields
+            rename_field(elem, "enabled", "is_active")
+            rename_field(elem, "id", "c4_id")
+            rename_field(elem, "phone_prefix", "c4_phone_prefix")
+            rename_field(elem, "time_zone", "c4_time_zone")
+            rename_field(elem, "website", "c4_website")
+            rename_field(elem, "company_name", "c4_company_name")
+            rename_field(elem, "siret", "c4_siret")
+            rename_field(elem, "naf", "c4_naf")
+            rename_field(elem, "phone_verified", "c4_phone_verified")
+            rename_field(elem, "email_verified", "c4_email_verified")
+            rename_field(elem, "id_card_verified", "c4_id_card_verified")
+            rename_field(elem, "accept_survey", "c4_accept_survey")
+            rename_field(elem, "accept_rgpd", "c4_accept_rgpd")
+            rename_field(elem, "offers_for_pro_sector", "c4_offers_for_pro_sector")
+            rename_field(elem, "quote_promise", "c4_quote_promise")
+
+            # cleanup fields
+            cleanup_date_field_names(elem)
+            make_aware_dates(elem)
+            integer_to_boolean(elem)
+
+            # cleanup person_type
+            if "person_type" in elem:
+                elem["kind"] = map_user_kind(elem["person_type"])
+
+            # set staff users
+            if "roles" in elem:
+                if elem["roles"].startswith("a:1:{i:0;s:10"):
+                    elem["is_staff"] = True
+                if elem["roles"].startswith("a:1:{i:0;s:16"):
+                    elem["is_superuser"] = True
+
+            # remove useless keys
+            [elem.pop(key) for key in USER_EXTRA_KEYS]
+
+            # create object
+            # Note: we ignore users with kind=None
+            if elem["kind"]:
+                try:
+                    first = User.objects.create(**elem)  # noqa
+                except Exception as e:
+                    print("a", e)
+
+        print(f"Created {User.objects.count()} users !")
+
+    def migrate_siae_user(self, cur):
+        """
+        Migrate M2M data between Siae & User
+
+        Notes:
+        - elem exemple: {'directory_id': 270, 'user_id': 471234844}
+        """
+        print("Migrating M2M between Siae & User...")
+
+        Siae.users.through.objects.all().delete()
+
+        cur.execute("SELECT * FROM directory_user")
+        resp = cur.fetchall()
+        # print(len(resp))
+        # print(resp[0])
+
+        for elem in resp:
+            try:
+                user = User.objects.get(c4_id=elem["user_id"])
+                user.siaes.add(elem["directory_id"])
+            # Note: some users were ignored because of kind=None. So we ignore the relation as well.
+            except:  # noqa
+                pass
+
+        print(f"Created {Siae.users.through.objects.count()} M2M objects !")