+++ CHANGELOG.md
@@ -10,10 +10,14 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 
 ### [0.8] - (2021-09)
 
-- Premier script de migration pour récupérer la donnée des Siae, Réseaux, Secteurs, Prestations et Labels
+- Premier script de migration pour récupérer la donnée des Siae, Réseaux, Secteurs, Prestations, Labels et Références
 - Ajout du modèle SiaeOffer
 - Ajout du modèle SiaeLabel
-- API : la données des Sector provient des nouveaux modèles
+- Ajout du modèle SiaeClientReference
+- Ajout de Flatpages pour créer des pages statiques directement dans l'admin
+- API : les données des Sector proviennent du nouveau modèle
+- Outils : ajout des packets django-debug-toolbar & django-extensions
+- Outils : ajout d'un template de PR
 
 ### [0.7] - (2021-08)